Rename rtc.rs to time_driver.rs
We will add another time driver for user selection.
